requirements.txt
asgiref==3.7.2
asttokens==2.4.1
black==24.1.1
click==8.1.7
decorator==5.1.1
Django==4.2.9
djangorestframework==3.14.0
exceptiongroup==1.2.0
executing==2.0.1
ipython==8.20.0
jedi==0.19.1
matplotlib-inline==0.1.6
mypy-extensions==1.0.0
packaging==23.2
parso==0.8.3
pathspec==0.12.1
pexpect==4.9.0
platformdirs==4.1.0
prompt-toolkit==3.0.43
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.17.2
python-dotenv==1.0.1
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.3
tomli==2.0.1
traitlets==5.14.1
typing_extensions==4.9.0
wcwidth==0.2.13